+++ app/balance/route.ts
@@ -1,17 +1,10 @@
-import { getBalance, requireNpub } from "@/lib/server/auth";
+import { getBalance } from "@/lib/server/auth";
 
 export async function GET() {
-  try {
-    const npub = requireNpub();
+  const balance = await getBalance();
 
-    return Response.json({
-      success: true,
-      data: await getBalance(npub),
-    });
-  } catch (err) {
-    return Response.json({
-      success: false,
-      message: (err as Error).message,
-    });
-  }
+  return Response.json({
+    success: true,
+    data: balance,
+  });
 }
