+++ app/chat/index.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { useAppState } from "@/components/providers/app-state-provider";
 import Conversation from "./conversation";
 import EmptyState from "./empty";
