+++ app/page.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import Container from "@/components/container";
 import { AppStateProvider } from "@/components/providers/app-state-provider";
 import {
   NostrProvider,
@@ -12,30 +13,28 @@ import {
 import { Button } from "@/components/ui/button";
 import Icon from "@/components/ui/icon";
 import { Text } from "@/components/ui/text";
-import { useState } from "react";
 import Chat from "./chat";
+import TopupDialog from "./topup";
 
 function Fallback() {
   const { isLoading, error } = useNDKContext();
   const { isLoading: weblnLoading, error: weblnError } = useWebLNContext();
 
-  const [open, setOpen] = useState(false);
-
   if (isLoading || weblnLoading) {
     return (
-      <div className="flex items-center justify-center grow w-full h-full max-w-[480px]">
+      <Container center>
         <Icon
           icon="IconLoader2"
           size="xl"
           className="animate-load text-lightGrey"
         />
-      </div>
+      </Container>
     );
   }
 
   if (error) {
     return (
-      <div className="flex flex-col gap-lg items-center justify-center grow w-full h-full max-w-[480px]">
+      <Container center>
         <Text variant="h2" weight="bold" className="text-center">
           Nostr Provider Required
         </Text>
@@ -43,13 +42,13 @@ function Fallback() {
           A Nostr Provider is required in order to run this application.
         </Text>
         <Button onClick={() => window.location.reload()}>Reload</Button>
-      </div>
+      </Container>
     );
   }
 
   if (weblnError) {
     return (
-      <div className="flex flex-col gap-lg items-center justify-center grow w-full h-full max-w-[480px]">
+      <Container center>
         <Text variant="h2" weight="bold" className="text-center">
           WebLN Provider Required
         </Text>
@@ -57,17 +56,16 @@ function Fallback() {
           A WebLN Provider is required in order to run this application.
         </Text>
         <Button onClick={() => window.location.reload()}>Reload</Button>
-      </div>
+      </Container>
     );
   }
 
   return (
     <AppStateProvider>
-      <div className="w-full flex flex-col items-center justify-center min-h-screen">
-        <div className="flex flex-col grow w-full h-full max-w-[480px] divide-y divide-extraLightGrey">
-          <Chat />
-        </div>
-      </div>
+      <Container className="divide-y divide-y-extraLightGrey items-stretch">
+        <Chat />
+      </Container>
+      <TopupDialog />
     </AppStateProvider>
   );
 }
